Added muted sections detection
Since twitch only returns playlist with unmuted stream, but prohibits to download the "closed" segments, the playlist is downloaded locally and modified ("-muted" is added to the segment filename).
Though the VOD JSON contains the list of muted sections, I discovered that the time markers in the playlist are wildly inaccurate and do not allow to determine the correct segment files to be muted. Therefore, mute check is performed by trying each segment file availability. If a segment file cannot be downloaded it is considered muted (and if the muted version also cannot be downloaded, an error is returned).
Because this check is performed at startup and can take long time (several minutes), additional progress report was added.
Known issue: for highlights, the playlist still reports the full time length, so the analyzing progress will stop before reaching the designated end (highlight is ended, but is shorter than the full video).