Got rid of startup analyze stage
Looks like time markers are not incorrect, after all, it's just that muting works on whole-file basis, but mute segments are specified independently of the file segments. And if a file segment intersects a mute segment, the whole file becomes muted. Now this algorithm is taken into account, and availability-check is no longer required.